--- src/main/java/com/example/soen341/AnswerRepository.java
@@ -1,7 +1,0 @@
-package com.example.soen341;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface AnswerRepository
-        extends MongoRepository<Answer, String> {
-}
