+++ src/main/java/com/example/soen341/Model/User.java
@@ -1,7 +1,8 @@
-package com.example.soen341;
+package com.example.soen341.Model;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDateTime;
@@ -12,7 +13,9 @@ public class User {
 
     @Id
     private String id;
+    @Indexed(unique = true)
     private String username;
+    @Indexed(unique = true)
     private String email;
     private String password;
     private LocalDateTime created;
