--- src/main/java/com/example/soen341/QuestionRepository.java
@@ -1,8 +1,0 @@
-package com.example.soen341;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface QuestionRepository
-        extends MongoRepository<Question, String> {
-}
-
