--- src/main/java/com/example/soen341/VoteRepository.java
@@ -1,7 +1,0 @@
-package com.example.soen341;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface VoteRepository
-        extends MongoRepository<Vote, String> {
-}
