+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.data.mongodb.auto-index-creation=true
+spring.data.mongodb.database =stack
